added ArrayForTrees
